fix(ci): preserve gateway boot compatibility
