+++ src/opensquilla/gateway/boot.py
@@ -4907,6 +4907,9 @@ async def start_gateway_server(
     register_channels_reconciler(_reconcile_runtime_channels)
 
     # ── ASGI app ─────────────────────────────────────────────────────
+    skill_management_state = getattr(svc, "skill_management_state", None)
+    if skill_management_state is None:
+        skill_management_state = {}
     app = create_gateway_app(
         config,
         session_manager=svc.session_manager,
@@ -4921,7 +4924,7 @@ async def start_gateway_server(
         meta_run_writer=getattr(svc, "meta_run_writer", None),
         skill_loader=svc.skill_loader,
         skill_management_service=getattr(svc, "skill_management_service", None),
-        skill_management_state=svc.skill_management_state,
+        skill_management_state=skill_management_state,
         cron_scheduler=svc.cron_scheduler,
         turn_runner=turn_runner,
         task_runtime=task_runtime,
